Fix missing final boundary when no attachments
The final boundary line was missing when there are no attachments. It should be included regardless, since the multipart has already been established.
Without this, the server doesn't know you are actually done when the \r\n.\r\n is sent, and a timeout will occur, since it doesn't send 250 OK.